requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.6
django-debug-toolbar==4.4.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
notifiers==1.3.3
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
PyJWT==2.8.0
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
